+++ packages/models/lib/src/widgets/nodes/node_overrides.dart
@@ -1,15 +1,16 @@
 import 'package:equatable/equatable.dart';
+import 'package:flutter/material.dart';
 
 enum NodeProperties {
   child,
   children,
 }
 
-class NodeOverride extends Equatable {
-  const NodeOverride(
-    this.nodeIdenfier,
-    this.properties,
-  );
+class Override extends Equatable {
+  const Override({
+    required this.nodeIdenfier,
+    required List<NodeProperty> props,
+  }) : properties = props;
 
   final String nodeIdenfier;
   final List<NodeProperty> properties;
@@ -22,7 +23,7 @@ class NodeOverride extends Equatable {
       'Action { nodeIdenfier: $nodeIdenfier, properties: $properties }';
 }
 
-class NodeProperty extends Equatable {
+abstract class NodeProperty extends Equatable {
   const NodeProperty(
     this.property,
     this.value,
@@ -30,6 +31,28 @@ class NodeProperty extends Equatable {
 
   final NodeProperties property;
   final dynamic value;
+
+  @override
+  List<Object> get props => [property, value];
+}
+
+class ChildProperty extends NodeProperty {
+  const ChildProperty({
+    required this.child,
+  }) : super(NodeProperties.child, child);
+
+  final Widget child;
+
+  @override
+  List<Object> get props => [property, value];
+}
+
+class ChildrenProperty extends NodeProperty {
+  const ChildrenProperty({
+    required this.children,
+  }) : super(NodeProperties.children, children);
+
+  final List<Widget> children;
 
   @override
   List<Object> get props => [property, value];